+++ src/Cli/SubCommands/index.ts
@@ -2,11 +2,13 @@ import GetVersionCommand from "./GetVersionCommand"
 import OpenUiCommand from "./OpenUiCommand"
 import HelpCommand from "./HelpCommand"
 import Command from "./Command"
+import StartServerCommand from "./StartServerCommand"
 
 const commands: Command[] = [
     GetVersionCommand,
-    OpenUiCommand,
     HelpCommand,
+    OpenUiCommand,
+    StartServerCommand,
 ]
 
 export default commands