+++ src/Server/run-server.ts
@@ -1,5 +1,15 @@
-import server from './server'
+import './Startup/preflight'
+import makeServer from './makeServer'
+import getDependencies from "./Dependencies/getDependencies"
 
-server.listen(8777, () => {
-    console.log('Listening on 8777')
-})
+async function runServer() {
+    const deps = await getDependencies()
+    const server = await makeServer(deps)
+
+    server.listen(8777, () => {
+        console.log('Listening on 8777')
+    })
+
+}
+
+runServer()