+++ src/Server/server.test.ts
@@ -1,7 +1,0 @@
-import server from './server'
-
-describe('server', () => {
-    it('should be truthy', () => {
-        expect(server).toBeTruthy()
-    })
-})